# order-management-api/src/main/api/openapi.yaml
openapi: 3.0.0
info:
  title: Order Management API
  version: '@VERSION@'
servers:
- url: https://localhost/v1
tags:
- name: orders
  description: Order Management
paths:
  /orders:
    get:
      tags:
      - orders
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/order'
      operationId: GetOrders
    post:
      tags:
      - orders
      responses:
        201:
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/order'
      operationId: PostOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/orderInput'
        required: true
  /orders/{orderId}:
    get:
      tags:
      - orders
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/order'
      operationId: GetOrderByOrderId
    put:
      tags:
      - orders
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/order'
      operationId: PutOrderByOrderId
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/orderInput'
        required: true
    delete:
      tags:
      - orders
      responses:
        200:
          description: ''
      operationId: DeleteOrderByOrderId
    parameters:
    - name: orderId
      in: path
      required: true
      schema:
        type: integer
        format: int64
components:
  schemas:
    order:
      type: object
      properties:
        id:
          type: integer
          format: int64
        version:
          type: integer
          format: int64
        price:
          type: number
        items:
          type: array
          items:
            $ref: '#/components/schemas/orderItem'
        createdOn:
          type: string
          format: date-time
        lastModified:
          type: string
          format: date-time
      required:
      - id
      - version
      - price
      - items
      - createdOn
      - lastModified
    orderItem:
      type: object
      properties:
        quantity:
          type: integer
          format: int32
        productId:
          type: string
        price:
          type: number
      required:
      - quantity
      - productId
      - price
    orderInput:
      type: object
      properties:
        version:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: '#/components/schemas/orderItemInput'
      required:
      - items
    orderItemInput:
      type: object
      properties:
        quantity:
          type: integer
          format: int32
        productId:
          type: string
      required:
      - quantity
      - productId
  responses: {}
  parameters: {}
  examples: {}
  requestBodies: {}
  headers: {}
  securitySchemes: {}
  links: {}
  callbacks: {}